hibakezeles

+++ haromszogek/Haromszog.cs
@@ -6,9 +6,9 @@ namespace haromszogek
 {
     internal class Haromszog
     {
-        private int aOldal;
-        private int bOldal;
-        private int cOldal;
+        private double aOldal;
+        private double bOldal;
+        private double cOldal;
 
         public double Terulet { get; private set; }
         public double Kerulet { get; private set; }
@@ -54,7 +54,7 @@ namespace haromszogek
             }
             return adatok;
         }
-        public Haromszog(int aOldal, int bOldal, int cOldal)
+        public Haromszog(double aOldal, double bOldal, double cOldal)
         {
             this.aOldal = aOldal;
             this.bOldal = bOldal;
